recevoir envoyer ok

+++ app/src/main/java/com/dut/kidoi/models/Envoyer.java
@@ -1,7 +1,7 @@
 package com.dut.kidoi.models;
 
 public class Envoyer extends Transaction {
-    public Envoyer(String username, String message, float montant, boolean fait) {
+    public Envoyer(String username, String message, Long montant, boolean fait) {
         super(username, message, montant, fait);
     }
